requirements.txt
Pillow==9.3.0
Pillow==10.2.0
